Factor out TestValue constructs from ResolverTest
